# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Uptime Kuma API Bridge",
    "version": "1.0.0",
    "description": "Private REST bridge to provision and synchronize Uptime Kuma monitors. Every request (except /health and /ready) requires `Authorization: Bearer <BRIDGE_AUTH_TOKEN>`."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bridge's own static token (BRIDGE_AUTH_TOKEN), unrelated to Uptime Kuma accounts."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/monitors": {
      "get": {
        "summary": "List monitors",
        "tags": [
          "Monitors"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "http",
                "group"
              ]
            },
            "in": "query",
            "name": "type",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "name": {
                        "type": "string"
                      },
                      "description": {
                        "nullable": true,
                        "type": "string"
                      },
                      "parent": {
                        "nullable": true,
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "http",
                          "group"
                        ]
                      },
                      "active": {
                        "type": "boolean"
                      },
                      "url": {
                        "nullable": true,
                        "type": "string"
                      },
                      "method": {
                        "type": "string"
                      },
                      "interval": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "retryInterval": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "resendInterval": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "maxretries": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "timeout": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "acceptedStatusCodes": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "maxRedirects": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "keyword": {
                        "nullable": true,
                        "type": "string"
                      },
                      "invertKeyword": {
                        "type": "boolean"
                      },
                      "ignoreTls": {
                        "type": "boolean"
                      },
                      "upsideDown": {
                        "type": "boolean"
                      },
                      "notificationIds": {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "tagIds": {
                        "type": "array",
                        "items": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "expiryNotification": {
                        "type": "boolean"
                      },
                      "domainExpiryNotification": {
                        "type": "boolean"
                      },
                      "cacheBust": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "description",
                      "parent",
                      "type",
                      "active"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an HTTP(S) monitor",
        "tags": [
          "Monitors"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "parent": {
                    "default": null,
                    "nullable": true,
                    "type": "integer",
                    "minimum": 0,
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991
                  },
                  "description": {
                    "default": null,
                    "nullable": true,
                    "type": "string"
                  },
                  "method": {
                    "default": "GET",
                    "type": "string",
                    "enum": [
                      "GET",
                      "POST",
                      "PUT",
                      "PATCH",
                      "DELETE",
                      "HEAD",
                      "OPTIONS"
                    ]
                  },
                  "interval": {
                    "default": 60,
                    "type": "integer",
                    "minimum": 20,
                    "maximum": 9007199254740991
                  },
                  "retryInterval": {
                    "type": "integer",
                    "minimum": 20,
                    "maximum": 9007199254740991
                  },
                  "resendInterval": {
                    "default": 0,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "retries": {
                    "default": 0,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "timeout": {
                    "default": 48,
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "acceptedStatusCodes": {
                    "default": [
                      "200-299"
                    ],
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "maxRedirects": {
                    "default": 10,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "keyword": {
                    "default": null,
                    "nullable": true,
                    "type": "string"
                  },
                  "invertKeyword": {
                    "default": false,
                    "type": "boolean"
                  },
                  "ignoreTls": {
                    "default": false,
                    "type": "boolean"
                  },
                  "upsideDown": {
                    "default": false,
                    "type": "boolean"
                  },
                  "body": {
                    "default": null,
                    "nullable": true,
                    "type": "string"
                  },
                  "headers": {
                    "default": null,
                    "nullable": true,
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "basicAuth": {
                    "default": null,
                    "nullable": true,
                    "type": "object",
                    "properties": {
                      "username": {
                        "type": "string"
                      },
                      "password": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "username",
                      "password"
                    ]
                  },
                  "bearerToken": {
                    "default": null,
                    "nullable": true,
                    "type": "string"
                  },
                  "notificationIds": {
                    "default": [],
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true,
                      "maximum": 9007199254740991
                    }
                  },
                  "tagIds": {
                    "default": [],
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "minimum": 0,
                      "exclusiveMinimum": true,
                      "maximum": 9007199254740991
                    }
                  },
                  "expiryNotification": {
                    "default": false,
                    "type": "boolean"
                  },
                  "domainExpiryNotification": {
                    "default": false,
                    "type": "boolean"
                  },
                  "cacheBust": {
                    "default": false,
                    "type": "boolean"
                  },
                  "active": {
                    "default": true,
                    "type": "boolean"
                  }
                },
                "required": [
                  "name",
                  "url"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "name": {
                      "type": "string"
                    },
                    "description": {
                      "nullable": true,
                      "type": "string"
                    },
                    "parent": {
                      "nullable": true,
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "type": {
                      "type": "string",
                      "enum": [
                        "http",
                        "group"
                      ]
                    },
                    "active": {
                      "type": "boolean"
                    },
                    "url": {
                      "nullable": true,
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "interval": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "retryInterval": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "resendInterval": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "maxretries": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "timeout": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "acceptedStatusCodes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "maxRedirects": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "keyword": {
                      "nullable": true,
                      "type": "string"
                    },
                    "invertKeyword": {
                      "type": "boolean"
                    },
                    "ignoreTls": {
                      "type": "boolean"
                    },
                    "upsideDown": {
                      "type": "boolean"
                    },
                    "notificationIds": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      }
                    },
                    "tagIds": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      }
                    },
                    "expiryNotification": {
                      "type": "boolean"
                    },
                    "domainExpiryNotification": {
                      "type": "boolean"
                    },
                    "cacheBust": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "description",
                    "parent",
                    "type",
                    "active"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/monitors/{id}/pause": {
      "patch": {
        "summary": "Pause a monitor",
        "tags": [
          "Monitors"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/monitors/{id}/resume": {
      "patch": {
        "summary": "Resume a monitor",
        "tags": [
          "Monitors"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/monitors/{id}/heartbeats": {
      "get": {
        "summary": "Raw heartbeat history for a monitor",
        "tags": [
          "Monitors"
        ],
        "parameters": [
          {
            "schema": {
              "default": 24,
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "period",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "monitorId": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "status": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "time": {
                        "type": "string"
                      },
                      "msg": {
                        "nullable": true,
                        "type": "string"
                      },
                      "important": {
                        "type": "boolean"
                      },
                      "duration": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "ping": {
                        "nullable": true,
                        "type": "number"
                      },
                      "downCount": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      }
                    },
                    "required": [
                      "id",
                      "monitorId",
                      "status",
                      "time",
                      "msg",
                      "important",
                      "duration",
                      "ping",
                      "downCount"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/monitors/{id}/chart": {
      "get": {
        "summary": "Bucketed up/down/ping chart data for a monitor",
        "tags": [
          "Monitors"
        ],
        "parameters": [
          {
            "schema": {
              "default": 24,
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "period",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "timestamp": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "up": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "down": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      },
                      "avgPing": {
                        "nullable": true,
                        "type": "number"
                      },
                      "minPing": {
                        "nullable": true,
                        "type": "number"
                      },
                      "maxPing": {
                        "nullable": true,
                        "type": "number"
                      }
                    },
                    "required": [
                      "timestamp",
                      "up",
                      "down",
                      "avgPing",
                      "minPing",
                      "maxPing"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/monitors/{id}/stats": {
      "get": {
        "summary": "Uptime %, average ping, and cert/domain expiry for a monitor",
        "tags": [
          "Monitors"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 9007199254740991
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "uptime24h": {
                      "nullable": true,
                      "type": "number"
                    },
                    "uptime30d": {
                      "nullable": true,
                      "type": "number"
                    },
                    "uptime1y": {
                      "nullable": true,
                      "type": "number"
                    },
                    "avgPing24h": {
                      "nullable": true,
                      "type": "number"
                    },
                    "certInfo": {
                      "nullable": true
                    },
                    "domainDaysRemaining": {
                      "nullable": true,
                      "type": "number"
                    },
                    "domainExpiresOn": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "required": [
                    "uptime24h",
                    "uptime30d",
                    "uptime1y",
                    "avgPing24h",
                    "certInfo",
                    "domainDaysRemaining",
                    "domainExpiresOn"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:3050"
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}